11/10

+++ src/Posicion.java
@@ -1,11 +1,11 @@
 
 public class Posicion {
 
-	private double x;
-	private double y;
+	private int x;
+	private int y;
 	
 	
-	public Posicion (double x, double y){
+	public Posicion (int x, int y){
 		this.x = x;
 		this.y = y; 
 	}
@@ -18,19 +18,19 @@ public class Posicion {
 		return(this.getX() <= elemento.getX() && this.getY() <= elemento.getY());
 	}
 
-	public double getX() {
+	public int getX() {
 		return x;
 	}
 
-	public void setX(double x) {
+	public void setX(int x) {
 		this.x = x;
 	}
 
-	public double getY() {
+	public int getY() {
 		return y;
 	}
 
-	public void setY(double y) {
+	public void setY(int y) {
 		this.y = y;
 	}
